+++ src/fl_draw_image_win32.cxx
@@ -1,5 +1,5 @@
 //
-// "$Id: fl_draw_image_win32.cxx,v 1.4.2.3.2.2 2002/01/01 15:11:32 easysw Exp $"
+// "$Id: fl_draw_image_win32.cxx,v 1.4.2.3.2.3 2002/11/19 16:37:36 easysw Exp $"
 //
 // WIN32 image drawing code for the Fast Light Tool Kit (FLTK).
 //
@@ -137,17 +137,17 @@ static void innards(const uchar *buf, int X, int Y, int W, int H,
   }
 #if USE_COLORMAP
   if (indexed) {
-    for (int i=0; i<256; i++) {
+    for (short i=0; i<256; i++) {
       *((short*)(bmi.bmiColors)+i) = i;
     }
   } else
 #endif
   if (mono) {
     for (int i=0; i<256; i++) {
-      bmi.bmiColors[i].rgbBlue = i;
-      bmi.bmiColors[i].rgbGreen = i;
-      bmi.bmiColors[i].rgbRed = i;
-      bmi.bmiColors[i].rgbReserved = i;
+      bmi.bmiColors[i].rgbBlue = (uchar)(i;
+      bmi.bmiColors[i].rgbGreen = (uchar)(i;
+      bmi.bmiColors[i].rgbRed = (uchar)(i;
+      bmi.bmiColors[i].rgbReserved = (uchar)(i;
     }
   }
   bmi.bmiHeader.biWidth = w;
@@ -258,5 +258,5 @@ void fl_rectf(int x, int y, int w, int h, uchar r, uchar g, uchar b) {
 }
 
 //
-// End of "$Id: fl_draw_image_win32.cxx,v 1.4.2.3.2.2 2002/01/01 15:11:32 easysw Exp $".
+// End of "$Id: fl_draw_image_win32.cxx,v 1.4.2.3.2.3 2002/11/19 16:37:36 easysw Exp $".
 //
